requirements.txt
Django==1.6.4
Jinja2==2.7.2
MarkupSafe==0.23
Pygments==1.6
Sphinx==1.2.2
dj-database-url==0.3.0
dj-static==0.0.5
django-allauth==0.16.1
django-filter==0.7
django-crispy-forms==1.4.0
django-toolbelt==0.0.1
djangorestframework==2.3.14
docutils==0.11
gunicorn==18.0
markdown==2.4.1
mock==1.0.1
model-mommy==1.2.1
oauthlib==0.6.1
psycopg2==2.5.2
pystache==0.5.3
python-openid==2.2.5
requests==2.3.0
requests-oauthlib==0.4.0
six==1.3.0
static==1.0.2
wsgiref==0.1.2
